feat(web-cloud.emailpro): emailpro and exchange derives from autoconf
  ref: MANAGER-12343
