+++ packages/manager/modules/exchange/src/domain/dkim-autoconfig/domain-dkim-autoconfig.controller.js
@@ -1,6 +1,8 @@
 import { DKIM_CONFIGURATION_GUIDE } from './domain-dkim-autoconfig.constants';
 
-export default class ExchangeDomainDkimAutoconfigCtrl {
+import DkimAutoConfigurator from './dkim-auto-configurator';
+
+export default class ExchangeDomainDkimAutoconfigCtrl extends DkimAutoConfigurator {
   /* @ngInject */
   constructor(
     $q,
@@ -13,6 +15,7 @@ export default class ExchangeDomainDkimAutoconfigCtrl {
     exchangeStates,
     coreConfig,
   ) {
+    super();
     this.services = {
       $q,
       $scope,
@@ -35,6 +38,8 @@ export default class ExchangeDomainDkimAutoconfigCtrl {
       DKIM_CONFIGURATION_GUIDE.DEFAULT;
 
     this.loading = true;
+
+    this.initializeScope($scope);
 
     // Vars for DKIM configuration inside modal stepper
     this.initializeDkimConfiguratorNoOvh();
@@ -65,114 +70,7 @@ export default class ExchangeDomainDkimAutoconfigCtrl {
         );
       });
 
-    $scope.onFinishDkim = () => this.onFinishDkim();
-
-    // DKIM configurator methods for no Ovh Cloud
-    $scope.loadDataForDkim = () => this.loadDataForDkim();
-    $scope.getDkimSelector = () => this.getDkimSelector();
-    $scope.leaveDkimConfigurator = () => this.leaveDkimConfigurator();
-    $scope.getNextButtonDependingOnStep = () =>
-      this.getNextButtonDependingOnStep();
-    $scope.getTitleDependingOnStep = () => this.getTitleDependingOnStep();
-    $scope.initSpfContext = () => this.initSpfContext();
-  }
-
-  initializeDkimConfiguratorNoOvh() {
-    this.selector1NameInfos = '';
-    this.selector1RecordInfos = '';
-    this.selector2NameInfos = '';
-    this.selector2RecordInfos = '';
-    this.dkimSelectorsNoDomain = null;
-    this.selector1NoDomain = null;
-    this.selector2NoDomain = null;
-    this.isStepConfigureValid = false;
-    this.showConfiguratingBtn = true;
-    this.showSpfDiagnosticTitle = false;
-  }
-
-  hideConfirmButton() {
-    const showWhen = [
-      this.GLOBAL_DKIM_STATUS.NOT_CONFIGURED,
-      this.GLOBAL_DKIM_STATUS.OK,
-      this.GLOBAL_DKIM_STATUS.DISABLED,
-    ];
-
-    return this.domainDiag.isOvhDomain && showWhen.includes(this.dkimStatus);
-  }
-
-  onFinishDkim() {
-    switch (this.dkimStatus) {
-      case this.GLOBAL_DKIM_STATUS.NOT_CONFIGURED:
-        return this.configDkim();
-      case this.GLOBAL_DKIM_STATUS.OK:
-        return this.deactivateDkim();
-      case this.GLOBAL_DKIM_STATUS.DISABLED:
-        return this.activateDkim();
-      default:
-        console.error('Invalid DKIM status:', this.dkimStatus);
-        return null;
-    }
-  }
-
-  activateDkim() {
-    const dkimSelectors = this.domain.dkim;
-
-    this.services.ExchangeDomains.enableDkim(
-      this.$routerParams.organization,
-      this.$routerParams.productId,
-      this.domain.name,
-      dkimSelectors[0].selectorName,
-    ).finally(() => {
-      this.services.navigation.resetAction();
-    });
-  }
-
-  deactivateDkim() {
-    const dkimSelectors = this.domain.dkim;
-
-    this.services.ExchangeDomains.disableDkim(
-      this.$routerParams.organization,
-      this.$routerParams.productId,
-      this.domain.name,
-      dkimSelectors[0].status === this.DKIM_STATUS.IN_PRODUCTION
-        ? dkimSelectors[0].selectorName
-        : dkimSelectors[1].selectorName,
-    ).finally(() => {
-      this.services.navigation.resetAction();
-    });
-  }
-
-  async getSelectorNameForNoOvhCloud() {
-    if (!this.dkimSelectorsNoDomain) {
-      return;
-    }
-    const promises = this.dkimSelectorsNoDomain.map((dkimSelector) => {
-      return this.services.ExchangeDomains.getDkimSelectorName(
-        this.$routerParams.organization,
-        this.$routerParams.productId,
-        this.domain.name,
-        dkimSelector,
-      );
-    });
-
-    const [selector1NoDomain, selector2NoDomain] = await this.services.$q.all(
-      promises,
-    );
-    this.selector1NoDomain = selector1NoDomain;
-    this.selector1NameInfos = this.getDkimName(1);
-    this.selector1RecordInfos = this.getDkimRecord(1);
-    this.selector2NoDomain = selector2NoDomain;
-    this.selector2NameInfos = this.getDkimName(2);
-    this.selector2RecordInfos = this.getDkimRecord(2);
-  }
-
-  async loadDataForDkim() {
-    this.showConfiguratingBtn = false;
-    this.showSpfDiagnosticTitle = false;
-    this.loading = true;
-    await this.stepConfigureDkim();
-    this.isStepConfigureValid = true;
-    this.loading = false;
+    $scope.configDkim = () => this.configDkim();
   }
 
   getDkimSelectorForCurrentState() {
@@ -181,17 +79,6 @@ export default class ExchangeDomainDkimAutoconfigCtrl {
       this.$routerParams.productId,
       this.domain.name,
     );
-  }
-
-  getDkimSelector() {
-    return this.getDkimSelectorForCurrentState().then((dkimSelectors) => {
-      this.dkimSelectorsNoDomain = dkimSelectors;
-    });
-  }
-
-  leaveDkimConfigurator() {
-    this.services.navigation.resetAction();
-    return this.initializeDkimConfiguratorNoOvh();
   }
 
   postDkimFor(selectors) {
@@ -210,61 +97,23 @@ export default class ExchangeDomainDkimAutoconfigCtrl {
   }
 
   stepConfigureDkim() {
-    const promises = this.postDkimFor(this.dkimSelectorsNoDomain);
-    return this.services.$q
-      .all(promises)
-      .then(() => {
-        this.services.messaging.writeSuccess(
-          this.services.$translate.instant(
-            'exchange_tab_domain_diagnostic_dkim_activation_success',
-          ),
-        );
-      })
-      .catch(() => {
-        this.leaveDkimConfigurator();
-        this.services.messaging.writeError(
-          this.services.$translate.instant(
-            'exchange_tab_domain_diagnostic_dkim_configurate_no_ovhcloud_failed',
-          ),
-        );
-      });
-  }
-
-  initSpfContext() {
-    this.showSpfDiagnosticTitle = true;
-    return this.getSelectorNameForNoOvhCloud();
+    this.stepConfigureDkimFor('exchange');
   }
 
   getTitleDependingOnStep() {
-    return this.showSpfDiagnosticTitle
-      ? 'exchange_tab_domain_diagnostic_dkim_configurate_no_ovhcloud'
-      : 'exchange_tab_domain_diagnostic_dkim_next_no_ovhcloud';
+    return this.getTitleDependingOnStepFor('exchange');
   }
 
   getNextButtonDependingOnStep() {
-    return this.showConfiguratingBtn
-      ? 'exchange_tab_domain_diagnostic_dkim_title_configuration'
-      : 'exchange_tab_domain_diagnostic_spf_title';
+    return this.getNextButtonDependingOnStepFor('exchange');
   }
 
   getDkimName(index) {
-    return [
-      this.services.$translate.instant(
-        'exchange_tab_domain_diagnostic_dkim_name',
-      ),
-      `${index}: `,
-      this[`selector${index}NoDomain`].customerRecord,
-    ].join('');
+    return this.getDkimNameFor('exchange', index);
   }
 
   getDkimRecord(index) {
-    return [
-      this.services.$translate.instant(
-        'exchange_tab_domain_diagnostic_dkim_record',
-      ),
-      `${index}: `,
-      this[`selector${index}NoDomain`].targetRecord,
-    ].join('');
+    return this.getDkimRecordFor('exchange', index);
   }
 
   configDkim() {
